[remove] Batch Defination & Async Loading of ECharts components & charts
