+++ package.json
@@ -37,6 +37,10 @@
             "./dist/component/*.d.ts",
             "./dist/chart/*",
             "./dist/chart/*.d.ts"
+        ],
+        "./chart": [
+            "./dist/chart.js",
+            "./dist/chart.d.ts"
         ]
     },
     "engines": {
@@ -80,9 +84,9 @@
         "prepare": "husky",
         "test": "lint-staged",
         "clean": "rimraf .parcel-cache/ dist/ docs/",
-        "preview": "npm run clean  &&  cd preview/  &&  parcel --dist-dir=../docs/preview/ --open",
+        "preview": "npm run clean  &&  cd preview/  &&  parcel --dist-dir=../docs/preview/",
         "pack-preview": "rimraf .parcel-cache/ docs/preview/  &&  cd preview/  &&  parcel build --public-url=. --dist-dir=../docs/preview/",
-        "pack-dist": "rm -rf dist/  &&  tsc  &&  microbundle --name EChartsJSX --globals lodash=_,echarts/core=echarts,web-utility=WebUtility",
+        "pack-dist": "rm -rf dist/  &&  tsc  &&  microbundle --name EChartsJSX --globals lodash=_,echarts/core=echarts,echarts/charts=echarts,echarts/components=echarts,web-utility=WebUtility",
         "pack-docs": "rm -rf docs/  &&  typedoc source/",
         "build": "npm run pack-dist  &&  npm run pack-docs  &&  npm run pack-preview",
         "start": "npm run pack-docs  &&  open-cli docs/index.html",
