+++ source/charts/bar.ts
