+++ source/charts/effect-scatter.ts
