+++ source/charts/gauge.ts
