+++ source/charts/parallel.ts
