+++ source/charts/pictorial-bar.ts
