+++ source/components/axis-pointer.ts
