+++ source/components/legend.ts
