+++ source/components/parallel.ts
