+++ source/components/visual-map.ts
