+++ source/index.ts
@@ -1,48 +1,48 @@
 export * from './utility';
 export * from './Proxy';
 export * from './Option';
-export * from './Chart';
+export * from './chart';
 
-export * from './component/grid';
-export * from './component/polar';
-export * from './component/geo';
-export * from './component/single-axis';
-export * from './component/parallel';
-export * from './component/calendar';
-export * from './component/graphic';
-export * from './component/toolbox';
-export * from './component/tooltip';
-export * from './component/axis-pointer';
-export * from './component/brush';
-export * from './component/title';
-export * from './component/timeline';
-export * from './component/legend';
-export * from './component/data-zoom';
-export * from './component/visual-map';
-export * from './component/x-axis';
-export * from './component/y-axis';
-export * from './component/angle-axis';
-export * from './component/radius-axis';
+export * from './components/grid';
+export * from './components/polar';
+export * from './components/geo';
+export * from './components/single-axis';
+export * from './components/parallel';
+export * from './components/calendar';
+export * from './components/graphic';
+export * from './components/toolbox';
+export * from './components/tooltip';
+export * from './components/axis-pointer';
+export * from './components/brush';
+export * from './components/title';
+export * from './components/timeline';
+export * from './components/legend';
+export * from './components/data-zoom';
+export * from './components/visual-map';
+export * from './components/x-axis';
+export * from './components/y-axis';
+export * from './components/angle-axis';
+export * from './components/radius-axis';
 
-export * from './chart/line';
-export * from './chart/bar';
-export * from './chart/pie';
-export * from './chart/scatter';
-export * from './chart/radar';
-export * from './chart/map';
-export * from './chart/tree';
-export * from './chart/treemap';
-export * from './chart/graph';
-export * from './chart/gauge';
-export * from './chart/funnel';
-export * from './chart/parallel';
-export * from './chart/sankey';
-export * from './chart/boxplot';
-export * from './chart/candlestick';
-export * from './chart/effect-scatter';
-export * from './chart/lines';
-export * from './chart/heatmap';
-export * from './chart/pictorial-bar';
-export * from './chart/theme-river';
-export * from './chart/sunburst';
-export * from './chart/custom';
+export * from './charts/line';
+export * from './charts/bar';
+export * from './charts/pie';
+export * from './charts/scatter';
+export * from './charts/radar';
+export * from './charts/map';
+export * from './charts/tree';
+export * from './charts/treemap';
+export * from './charts/graph';
+export * from './charts/gauge';
+export * from './charts/funnel';
+export * from './charts/parallel';
+export * from './charts/sankey';
+export * from './charts/boxplot';
+export * from './charts/candlestick';
+export * from './charts/effect-scatter';
+export * from './charts/lines';
+export * from './charts/heatmap';
+export * from './charts/pictorial-bar';
+export * from './charts/theme-river';
+export * from './charts/sunburst';
+export * from './charts/custom';
